Added a check to assure that a temporary file is empty before trying to test with it.

+++ src/test.sh
@@ -139,6 +139,16 @@ run_file_test()
 
     umask 077
     file=`mktemp`
+
+    if [ -s "$file" ]
+    then
+        printf '\033[2K\015' 1>&2
+        printf '[ERROR] Temporary file is non-empty.\n' 1>&2
+        printf 'Hint: %s\n' "$hint" 1>&2
+        printf 'Source: %s\n' "$text" 1>&2
+        exit 1
+    fi
+
     printf '%s\n' "$text" > "$file"
     actual_output=`$PROGRAM -f "$file"`
     actual_code=$?
